requirements.txt
numpy>=1.20.0
spatialmath-python>=0.11.0
roboticstoolbox-python>=0.11.0
pytransform3d>=1.8.0
matplotlib>=3.4.0
swift-sim>=0.1.0 